UnsafeCell<Bump> to UnsafeCellWith<Bump>
Original intent was to satisfy clippy as this replaces
code that used &self to return a &mut Bump with a `with_mut` method.

Idea comes from Tokio. Wrap the UnsafeCell<T> in a new type that
provides `with` and a `with_mut` methods. The scope of the unsafe code
might be more clear at the call site.

The dereference of the pointer, the *const T, or the *mut T, still has
to be wrapped in `unsafe` but it may make it more clear to the caller
what safety concerns they are taking into their own hands: whether it
is just that they can guarantee exclusive access to the Bump, or whether
they also should be guaranteeing the lifetime of the allocation won't
exceed the Bump being reset.