Add a helper macro for properties & async
The to_owned macro is helpful to handle hooks with async components, but
when using properties it can be problematic because `cx` will be moved
in the async block, and will result in a compilation error. This adds a
new helper that allows to make either hooks or properties owned.