feat(cli): Generate icons for Android when building APK When building an APK, if an icon is specified in the configuration, it will generate `*dpi` variants to override the default Android icon. For simple SVG files, it additionally generates Android VectorDrawable XML for `anydpi`.
The icon file is specified in the following configuration:

```toml
[bundle]
icon = ["path/to/icon"]
```

However, there are two issues:

First, I'm not sure why the icon is defined as `Vec<String>`. I couldn't find any usage of this field in the project as a reference, so I default to using the first element as the icon.

Second, Android appears to apply some scaling to the foreground of Android VectorDrawable XML, causing the actual icon to be enlarged and cropped. This requires adding some margin when creating the SVG. Additionally, currently the same icon file is used for both foreground and background, and only SVGs composed of simple paths are supported.