Fix unsafety in `ReadableExt::deref_impl`
Current implementation transmutes the lifetime of `&self` into an arbitrary lifetime `'a`, which can be used to create a dangling reference to `self`. Currently not triggered as all call sites for `deref_impl` then restrict the lifetime back down to that of the `Deref` trait's borrow of `self`.

Using the new precise capturing syntax, `use<...>`, we can inform the compiler that the closure returned by `deref_impl` captures `Self` (and any inherit lifetimes that type may have), allowing it to precisely have the lifetime of `&self` as it should.

Also moved safety checks into a const context, allowing the assumption this technique relies on to be partially validated at compile time. The compiler may already optimise out the original `assert_eq`, but this improves DX by allowing faults in the assumption to be found at compile time rather than through tests.

With these changes, `deref_impl` is now safe to call, as all invariants it previous relied on the caller to uphold are now done so in the type system.

Changes require increasing MSRV to 1.87.0