Revert MSRV change by decoupling implementation details of `ReadableExt` from `Deref`
Type parameters of the implementation for `Readable/Ext` can leak into the `impl Fn` type returned by `deref_impl`. By extracting it into a discreet function accepting a type implementing `Readable`, we can assert that only the type `T` should be captured in the opaque type.

This is an alternative to `use<...>` precise capturing