install.sh: Change shell interpreter from sh to bash
The syntax on line 16 and elsewhere, i.e. `if [[ -t 1 ]]; then` doesn't work in sh - it's a bash feature.

Tested on Ubuntu 25.10. The current install script errors with:

```
./install.sh: 16: [[: not found
./install.sh: 50: [[: not found
./install.sh: 69: [[: not found
#=O#-    #       #                                                                                                                                                                                                                    curl: (22) The requested URL returned error: 404
```

Changing to `#!/bin/bash` solves the problem.

It was probably working in many places because some systems symlink /bin/sh to /bin/bash.